move usage to own pkg
